+++ src/configure.js
@@ -1,12 +1,9 @@
 import warning from './utils/warning';
-import DefaultTheme from './DefaultTheme';
+import {parsePath, addLeadingSlash} from './utils/path';
+import {DefaultTheme, DefaultResponsiveSizes} from './DefaultTheme';
 import specimens from './specimens';
 import requireModuleDefault from './utils/requireModuleDefault';
 import NotFound from './components/Page/NotFound';
-
-// Removes potential multiple slashes from concatenating paths
-const removeMultiSlashes = (path) => path.replace(/\/+/g, '/');
-const stripTrailingSlashes = (path) => path.replace(/\/+$/, '');
 
 const has = (key) => (o) => o.hasOwnProperty(key);
 const hasName = has('name');
@@ -24,7 +21,7 @@ const flattenPageTree = (pageTree) => {
 
 export default (config) => {
   let pageId = 0;
-  const basePath = config.basePath || '/';
+  const basePath = addLeadingSlash(config.basePath || '/');
 
   const pageReducer = (pages, page) => {
     const configStyles = config.styles || [];
@@ -69,7 +66,7 @@ export default (config) => {
         ...page,
         id: ++pageId,
         // Currently, catalog can't be nested inside other page routes, it messes up <Link> matching. Use `basePath`
-        path: removeMultiSlashes('/' + stripTrailingSlashes([basePath, page.path || page.name].join('/'))),
+        path: parsePath(page.path || page.name, {basePath}).pathname,
         pages: page.pages ? page.pages.reduce(pageReducer, []).map((p) => ({...p, superTitle: page.title})) : null,
         styles: Array.from(new Set([...configStyles, ...pageStyles])),
         scripts: Array.from(new Set([...configScripts, ...pageScripts])),
@@ -80,7 +77,7 @@ export default (config) => {
 
   const pageTree = config.pages.reduce(pageReducer, []).map((p) => ({...p, superTitle: config.title}))
     .concat({
-      path: removeMultiSlashes('/' + stripTrailingSlashes([basePath, '*'].join('/'))),
+      path: parsePath('/*', {basePath}).pathname,
       id: ++pageId,
       component: NotFound,
       title: 'Page Not Found',
@@ -97,6 +94,7 @@ export default (config) => {
     // Used to check in configureRoutes() if input is already configured
     __catalogConfig: true,
     theme: {...DefaultTheme, ...config.theme},
+    responsiveSizes: config.responsiveSizes || DefaultResponsiveSizes,
     specimens: {...specimens, ...config.specimens},
     basePath,
     pages,