+++ src/specimens/Hint.js
@@ -81,7 +81,7 @@ class Hint extends React.Component {
             </div>
           </section>
       </div>
-      );
+    );
   }
 }